refactor: move serverPayload logic to a separate directory
