+++ packages/runtime/plugin-runtime/src/core/context/index.ts
@@ -75,7 +75,10 @@ interface GlobalContext {
 
 const globalContext: GlobalContext = {};
 
-export { getServerPayload, setServerPayload } from './serverPayload.server';
+export {
+  getServerPayload,
+  setServerPayload,
+} from './serverPayload/index';
 
 export function getGlobalIsRscClient() {
   return globalContext.isRscClient;
