+++ packages/runtime/plugin-runtime/src/core/context/serverPayload/index.tsx
@@ -1,0 +1,7 @@
+import type { ServerPayload } from '../index';
+
+export const getServerPayload = (): ServerPayload | undefined => {
+  return undefined;
+};
+
+export const setServerPayload = (payload: ServerPayload): void => {};
